refactor structure

+++ app/src/main/java/com/example/archivai/data/mappers/SectionMappers.kt
@@ -7,7 +7,7 @@ fun SectionResponseModel.toDomain(): Section {
     return Section(
         id = id,
         name = name,
-        size = size.toLong(),
+        size = size,
         lastModified = lastModified,
         numberOfFolders = numberOfFolders,
         numberOfEmployees = numberOfEmployees
